+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 #### Breaking Changes
 
-- EOL Ruby versions (2.3 and 2.4) are no longer supported
+- `ss.split(str, " ")` is no longer treated the same as `ss.split(str)` i.e.
+  unlike Ruby's `String#split` (but like Crystal's), the former no longer
+  strips the string before splitting
 - rename the `remove_empty` option `remove_empty_fields`
-- rename the `exclude` option (an alias for `reject`) `except`
+- rename the `exclude` option `except` (alias for `reject`)
 
 #### Fixes
 
@@ -13,15 +15,22 @@
 #### Features
 
 - add support for descending, negative, and infinite ranges,
-  e.g. `ss.split(str, ":", at: [..4, 4.., 3..1, -1..-3])` etc.
+  e.g. `ss.split(str, ":", at: [..4, 4..., 3..1, -1..-3])` etc.
 
 ## 0.5.1 - 2018-07-01
+
+#### Changes
 
 - set StringSplitter::VERSION when `string_splitter.rb` is loaded
 
 ## 0.5.0 - 2018-06-26
 
+#### Fixes
+
 - don't treat string delimiters as patterns
+
+#### Features
+
 - add a `reject`/`exclude` option which rejects splits at the specified positions
 - add a `select` alias for `at`
 
@@ -32,6 +41,8 @@
 - remove the `offset` alias for `split.index`
 
 ## 0.3.1 - 2018-06-24
+
+#### Fixes
 
 - remove trailing empty field when the separator is empty
   ([#1](https://github.com/chocolateboy/string_splitter/issues/1))